postgre sql

+++ src/main/java/tukano/impl/storage/CloudBlobStorage.java
@@ -1,14 +1,20 @@
 package tukano.impl.storage;
 
+import com.azure.core.http.rest.PagedIterable;
 import com.azure.core.util.BinaryData;
+import com.azure.core.util.Context;
 import com.azure.storage.blob.BlobClient;
+import com.azure.storage.blob.models.BlobItem;
+import com.azure.storage.blob.models.BlobRange;
 import com.azure.storage.blob.BlobClientBuilder;
 import com.azure.storage.blob.BlobContainerClient;
 import com.azure.storage.blob.BlobContainerClientBuilder;
+import com.azure.storage.blob.models.DownloadRetryOptions;
 import tukano.api.Result;
 import utils.Hash;
 import utils.IO;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -25,6 +31,7 @@ public class CloudBlobStorage implements BlobStorage {
 
     private final String storageConnectionString = "DefaultEndpointsProtocol=https;AccountName=sto62612northeurope;AccountKey=pYe+UhUDzJXoDsIzBG0SbAt4ic5yGQOwqQpFH0gjlpTDiD6RxiR6+VDUXXMyvEPHWBLJXwQKKAZy+AStoYUDiQ==;EndpointSuffix=core.windows.net";
     private static final String BLOBS_CONTAINER_NAME = "shorts";
+    private static final int CHUNK_SIZE = 4096;
     private static BlobContainerClient containerClient;
 
     public CloudBlobStorage() {
@@ -62,6 +69,9 @@ public class CloudBlobStorage implements BlobStorage {
 
         try {
             BlobClient blob = containerClient.getBlobClient(path);
+            if (!blob.exists())
+                return error(NOT_FOUND);
+
             blob.delete();
         } catch (Exception e) {
             e.printStackTrace();
@@ -79,7 +89,7 @@ public class CloudBlobStorage implements BlobStorage {
         try {
             BlobClient blob = containerClient.getBlobClient(path);
 
-            if (blob.exists())
+            if (!blob.exists())
                 return error(NOT_FOUND);
 
             bytes = blob.downloadContent();
@@ -92,6 +102,55 @@ public class CloudBlobStorage implements BlobStorage {
 
     @Override
     public Result<Void> read(String path, Consumer<byte[]> sink) {
-        return null;
+        if (path == null)
+            return error(BAD_REQUEST);
+
+        try {
+            BlobClient blob = containerClient.getBlobClient(path);
+            if (!blob.exists())
+                return error(NOT_FOUND);
+
+            long blobSize = blob.getProperties().getBlobSize();
+
+            for (long offset = 0; offset < blobSize; offset += CHUNK_SIZE) {
+                long chunkSize = (long) Math.min(CHUNK_SIZE, blobSize - offset);
+                ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+
+                /*blob.downloadStreamWithResponse(
+                        new ByteArrayOutputStream(),
+                        new BlobRange(offset, chunkSize),
+                        new DownloadRetryOptions().setMaxRetryRequests(5),
+                        null,
+                        false,
+                        new Context(key2, value2)).getStatusCode()
+                );*/
+                sink.accept(outputStream.toByteArray());
+            }
+            return ok();
+        } catch (Exception e) {
+            return error(INTERNAL_ERROR);
+        }
     }
-}
+
+    public Result<Void> deleteAll(String path) {
+        if (path == null)
+            return error(BAD_REQUEST);
+
+        try {
+
+            PagedIterable<BlobItem> allBlobs = containerClient.listBlobs();
+
+            for (BlobItem blobItem : allBlobs) {
+                BlobClient blob = containerClient.getBlobClient(blobItem.getName());
+                blob.delete();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return error(INTERNAL_ERROR);
+        }
+        return ok();
+    }
+
+
+    }
+
